Adding a way to establish a connection to the Intermec PC43d printer. In this commit we create 3 filed to handle that connection: socketconn.py, labelwriter.py,main.py.
In socketconn.py, we create a class that will handle connecting, sending, and receiving bytes of data to the printer. This is here for now so that we can send a beep to the printer once a connection is established to ensure that the printer received a command.

In labelwriter.py, we create a class that will take an ip address and port and establish a connection. This is a middle layer in between the websocket connection and the main program. We will use this to send the printer commands directly. Currently, the only command we are creating a function for is the beep to ensure that we can speak to the printer.

Lastly, main.py is where we will call the labelwriter with the commands we want to send it. We will use this as a place to make calls to the labelwriter class.